add check permission back, rename nav, some nav adjustments.
