+++ client/src/pages/Permits/style.js
@@ -5,9 +5,10 @@ import { NavLink } from 'react-router-dom';
 export const Layout = styled.nav`
     width:auto;
     padding: 14px;
-    margin-top: 12px;
     background-color: #FFFFFF;
     box-shadow: 0px 2px 13px rgba(0, 0, 0, 0.08);
+
+    margin: 12px -4.8rem 0 -4.8rem;
 `
 
 export const StyledLink = withTheme(styled(NavLink)`
